apps/workers/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision & ML
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1

# NLP & Language Processing
openai==1.3.7
transformers==4.36.2
spacy==3.7.2

# Database & Caching
asyncpg==0.29.0
redis==5.0.1
sqlalchemy[asyncio]==2.0.23

# Message Queue
nats-py==2.6.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
